// web/src/shared/api/schemas/projectsListResponseDto.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Translation API
 * The translation API description
 * OpenAPI spec version: 1.0
 */
import type { ProjectResponseDto } from './projectResponseDto';

export interface ProjectsListResponseDto {
  /** Массив проектов */
  data: ProjectResponseDto[];
  /** Общее количество проектов */
  total: number;
  /** Текущая страница */
  page: number;
  /** Количество проектов на странице */
  limit: number;
}